fixed getartworkcontroller, mismatch syntax for object type

+++ Application/src/main/java/com/kenzie/appserver/controller/GetArtworkController.java
@@ -41,7 +41,7 @@ public class GetArtworkController {
         artworkResponse.setTitle(artwork.getTitle());
         artworkResponse.setForSale(artwork.getIsForSale());
         artworkResponse.setSold(artwork.getIsSold());
-        artworkResponse.setPrice(artwork.getPrice().intValue());
+        artworkResponse.setPrice(artwork.getPrice());
 
         return artworkResponse;
     }
